Refactor substract function for both databases
