+++ index.js
@@ -5,7 +5,7 @@ const app = express();
 const bodyParser = require("body-parser");
 const messageValidated = require("./src/controllers/validations/messageValidations");
 const creditValidated = require("./src/controllers/validations/creditValidated");
-const createMessage = require("./src/database/saveMessage");
+const createMessage = require("./src/database/createMessage");
 const getMessages = require("./src/database/getMessages");
 const sendMessage = require("./src/controllers/sendMessage");
 const creditBalance = require("./src/controllers/creditBalance");
